Refactored to have simpler useEffect model
