+++ pages/consulting/index.tsx
@@ -1,11 +1,6 @@
-import classNames from "classnames";
-import Image from "next/image";
-import Link from "next/link";
-import { useRouter } from "next/router";
-import { useEffect, useMemo, useRef, useState } from "react";
-import { BsArrowRightCircle } from "react-icons/bs";
+import { NextRouter, useRouter } from "next/router";
+import { useEffect, useRef, useState } from "react";
 import { MdLiveHelp } from "react-icons/md";
-import { useHover } from "usehooks-ts";
 
 import { wrapGrid } from "animate-css-grid";
 
@@ -15,26 +10,13 @@ import { client } from "../../.tina/__generated__/client";
 import { InferGetStaticPropsType } from "next";
 import { ParsedUrlQuery } from "querystring";
 import { Breadcrumbs } from "../../components/blocks/breadcrumbs";
+import { Category } from "../../components/consulting/index/category";
+import { Tag } from "../../components/consulting/index/tag";
 import { Layout } from "../../components/layout";
 import { Container } from "../../components/util/container";
 import { SEO } from "../../components/util/seo";
 
 const allServices = "All SSW Services";
-
-const getSelectedTagFromQuery = (query: ParsedUrlQuery): string => {
-  let parsedTag = allServices;
-  if (query.tag) {
-    const { tag } = query;
-
-    if (tag instanceof Array) {
-      parsedTag = tag[0];
-    } else {
-      parsedTag = tag;
-    }
-    parsedTag = parsedTag.replace("-", " ");
-  }
-  return parsedTag;
-};
 
 export default function ConsultingIndex(
   props: InferGetStaticPropsType<typeof getStaticProps>
@@ -53,29 +35,11 @@ export default function ConsultingIndex(
   const [categories, setCategories] = useState([]);
   const [tags, setTags] = useState([]);
 
-  if (tags.some((x) => x.name === selectedTag)) {
-    router.push(
-      {
-        pathname: router.basePath,
-        query:
-          selectedTag === allServices
-            ? {}
-            : {
-                tag: selectedTag?.replace(" ", "-"),
-              },
-      },
-      undefined,
-      { shallow: true }
-    );
-  }
-
   useEffect(() => {
     // as the querystring changes, update the selected tag
     const qsTag = getSelectedTagFromQuery(router.query);
-    if (qsTag !== selectedTag) {
-      setSelectedTag(qsTag);
-    }
-  }, [router.query, selectedTag]);
+    setSelectedTag(qsTag);
+  }, [router.query]);
 
   useEffect(() => {
     // extract the data we need from the tina result
@@ -112,7 +76,10 @@ export default function ConsultingIndex(
                     label={tag.label}
                     tag={tag.name}
                     selectedTag={selectedTag}
-                    setSelectedTag={setSelectedTag}
+                    setSelectedTag={(val) => {
+                      updateParams(router, tags, val);
+                      setSelectedTag(val);
+                    }}
                   />
                 </div>
               ))}
@@ -140,116 +107,37 @@ export default function ConsultingIndex(
   );
 }
 
-const Tag = ({ label, tag, selectedTag, setSelectedTag }) => {
-  const isSelected = tag === selectedTag;
-  const hoverRef = useRef(null);
-  const hovered = useHover(hoverRef);
+const getSelectedTagFromQuery = (query: ParsedUrlQuery): string => {
+  let parsedTag = allServices;
+  if (query.tag) {
+    const { tag } = query;
 
-  return (
-    <li
-      className={classNames(
-        "cursor-pointer p-1 ease-in-out",
-        "duration-500 hover:bg-gray-100 hover:ease-in-out",
-        isSelected ? "text-sswRed" : ""
-      )}
-      onClick={() => setSelectedTag(tag)}
-      ref={hoverRef}
-    >
-      <div
-        className={classNames(
-          "inline-block h-3.5 w-6 text-sswRed",
-          !hovered && !isSelected && "-translate-x-2",
-          hovered && "translate-x-0 duration-300 ease-in"
-        )}
-      >
-        {(isSelected || hovered) && <BsArrowRightCircle />}
-      </div>
-      <span className="truncate">{label}</span>
-    </li>
-  );
+    if (tag instanceof Array) {
+      parsedTag = tag[0];
+    } else {
+      parsedTag = tag;
+    }
+    parsedTag = parsedTag.replace("-", " ");
+  }
+  return parsedTag;
 };
 
-const Category = ({ tinaData, category, selectedTag, index }) => {
-  const pages = useMemo(
-    () =>
-      category.pages.map((page) => {
-        return {
-          ...page,
-          isVisible: page.tags.includes(selectedTag),
-        };
-      }),
-    [category, selectedTag]
-  );
-
-  const categoryVisible = useMemo(
-    () => pages.some((page) => page.isVisible),
-    [pages]
-  );
-
-  const tinaCategory = tinaData.categories[index];
-
-  return (
-    <>
-      <div
-        className={classNames("lg:col-span-2", !categoryVisible && "hidden")}
-      >
-        {/* animate-css-grid requires a single element at this level */}
-        <div>
-          <hr className="my-5 border-gray-100" />
-          <h2
-            className="mt-0 text-sswRed"
-            data-tina-field={tinaField(tinaCategory.category, "name")}
-          >
-            {category.name}
-          </h2>
-        </div>
-      </div>
-      {pages.map((page, pageIndex) => (
-        <PageCard
-          key={page.title}
-          page={page}
-          category={tinaCategory}
-          pageIndex={pageIndex}
-        />
-      ))}
-    </>
-  );
-};
-
-const PageCard = ({ page, category, pageIndex }) => {
-  return (
-    <div
-      className={classNames(
-        "relative bg-white p-3 hover:bg-gray-50",
-        !page.isVisible && "hidden"
-      )}
-    >
-      {/* animate-css-grid requires a single element at this level */}
-      <div className="flex">
-        <div className="shrink-0">
-          {page.logo && (
-            <Image
-              className="mr-4 aspect-square h-14 w-14 border-1 border-gray-100 md:h-28 md:w-28"
-              height={115}
-              width={115}
-              src={page.logo}
-              alt={`${page.title} logo`}
-            />
-          )}
-        </div>
-        <div
-          className="min-w-0 flex-1"
-          data-tina-field={tinaField(category?.pages[pageIndex], "description")}
-        >
-          <Link href={page.url} className="unstyled">
-            <span className="absolute inset-0" aria-hidden="true" />
-            <h3 className="mb-2 mt-0 text-lg text-sswRed">{page.title}</h3>
-            <p className="text-sm text-black">{page.description}</p>
-          </Link>
-        </div>
-      </div>
-    </div>
-  );
+const updateParams = (router: NextRouter, tags, selectedTag) => {
+  if (tags.some((x) => x.name === selectedTag)) {
+    router.push(
+      {
+        pathname: router.basePath,
+        query:
+          selectedTag === allServices
+            ? {}
+            : {
+                tag: selectedTag?.replace(" ", "-"),
+              },
+      },
+      undefined,
+      { shallow: true }
+    );
+  }
 };
 
 const processData = (data) => {
